+++ tests/profiles/googlefonts_test.py
@@ -2087,6 +2087,11 @@ def test_check_metadata_os2_weightclass():
                            FAIL, 'mismatch',
                            f'with a bad metadata...')
 
+    ttFont = TTFont(TEST_FILE("leaguegothic-vf/LeagueGothic[wdth].ttf"))
+    assert_PASS(check(ttFont),
+                f'with a good VF that lacks a "wght" axis....')
+                # See: https://github.com/googlefonts/fontbakery/issues/3529
+
     # === test cases for Static Fonts ===
     # Our reference Montserrat family is a good 18-styles family:
     for fontfile in MONTSERRAT_RIBBI + MONTSERRAT_NON_RIBBI:
@@ -4076,3 +4081,21 @@ def test_check_description_urls():
     good_desc = check["description"].replace(">https://", ">")
     assert_PASS(check(font, {"description": good_desc}))
 
+
+def test_check_metadata_unsupported_subsets():
+    """Check for METADATA subsets with zero support."""
+    check = CheckTester(googlefonts_profile,
+                        "com.google.fonts/check/metadata/unsupported_subsets")
+
+    font = TEST_FILE("librecaslontext/LibreCaslonText[wght].ttf")
+    assert_PASS(check(font))
+
+    md = check["family_metadata"]
+    md.subsets.extend(["foo"])
+    assert_results_contain(check(font, {"family_metadata": md}),
+                           WARN, 'unknown-subset')
+
+    del md.subsets[:]
+    md.subsets.extend(["cyrillic"])
+    assert_results_contain(check(font, {"family_metadata": md}),
+                           WARN, 'unsupported-subset')