+++ CHANGELOG.md
@@ -11,6 +11,9 @@ A more detailed list of changes is available in the corresponding milestones for
 #### Added to the Universal Profile
   - **[com.google.fonts/check/cjk_chws_feature]:** Ensure CJK fonts contain chws/vchw features. (issue #3363)
 
+#### Added to the Google Fonts Profile
+  - **[com.google.fonts/check/metadata/unsupported_subsets]:** Check for METADATA subsets with zero support. (issue #3533)
+
 ### Changes to existing checks
 #### On the Universal Profile
   - **[com.google.fonts/check/unreachable_glyphs]:** Glyphs which are components of other glyphs are no longer flagged as unreachable. (issue #3523)